requirements.txt
asgiref==3.4.1
cachetools
certifi==2021.10.8
chardet==4.0.0
Django==3.2.11
django-phonenumber-field==6.0.0
google-api-core==2.4.0
google-auth
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-recaptcha-enterprise==1.5.0
google-cloud-secret-manager==2.8.0
googleapis-common-protos==1.54.0
grpc-google-iam-v1==0.12.3
grpcio==1.43.0
httplib2==0.20.2
idna==3.3
jsonfield==3.1.0
libcst==0.4.0
mypy-extensions==0.4.3
oauthlib==3.1.1
packaging==21.3
phonenumbers==8.12.41
proto-plus==1.19.8
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
pyparsing==3.0.6
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
sqlparse==0.4.2
stripe==2.64.0
twilio==7.5.0
typing-extensions==4.0.1
typing-inspect==0.7.1
urllib3==1.26.8
django-crispy-forms==1.13.0
sentry-sdk==1.5.11